change: gamebox를 함수형 컴포넌트로 전환
